requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-APScheduler==1.13.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
tenacity==8.2.3
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
